+++ packages/zowe-explorer/src/uss/actions.ts
@@ -133,7 +133,7 @@ export function deleteFromDisk(node: IZoweUSSTreeNode, filePath: string): void {
     }
 }
 
-export async function uploadDialog(node: IZoweUSSTreeNode, ussFileProvider: IZoweTree<IZoweUSSTreeNode>): Promise<void> {
+export async function uploadDialog(node: IZoweUSSTreeNode, ussFileProvider: IZoweTree<IZoweUSSTreeNode>, isBinary?: boolean): Promise<void> {
     ZoweLogger.trace("uss.actions.uploadDialog called.");
     const fileOpenOptions = {
         canSelectFiles: true,
@@ -143,20 +143,37 @@ export async function uploadDialog(node: IZoweUSSTreeNode, ussFileProvider: IZow
     };
 
     const value = await Gui.showOpenDialog(fileOpenOptions);
-
-    await Promise.all(
-        value.map(async (item) => {
-            const isBinary = isBinaryFileSync(item.fsPath);
-
-            if (isBinary) {
-                await uploadBinaryFile(node, item.fsPath);
-            } else {
-                const doc = await vscode.workspace.openTextDocument(item);
-                await uploadFile(node, doc);
+    if (value?.length > 0) {
+        await Gui.withProgress(
+            {
+                location: vscode.ProgressLocation.Notification,
+                title: localize("uploadFile.response.upload.title", "Uploading file to USS tree"),
+                cancellable: true,
+            },
+            async (progress, token) => {
+                let index = 0;
+                for (const item of value) {
+                    if (token.isCancellationRequested) {
+                        Gui.showMessage(localize("uploadFile.uploadCancelled", "Upload action was cancelled."));
+                        break;
+                    }
+                    Gui.reportProgress(progress, value.length, index, "Uploading");
+                    const binaryFile = isBinaryFileSync(item.fsPath);
+                    if (isBinary || binaryFile) {
+                        await uploadBinaryFile(node, item.fsPath);
+                    } else {
+                        const doc = await vscode.workspace.openTextDocument(item);
+                        await uploadFile(node, doc);
+                    }
+                }
+                index++;
             }
-        })
-    );
-    ussFileProvider.refresh();
+        );
+        ussFileProvider.refreshElement(node);
+        ussFileProvider.getTreeView().reveal(node, { expand: true, focus: true });
+    } else {
+        Gui.showMessage(localize("uploadFile.operationCancelled", "Operation Cancelled"));
+    }
 }
 
 export async function uploadBinaryFile(node: IZoweUSSTreeNode, filePath: string): Promise<void> {