+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentToolBar/PersistentClearAllButton.tsx
@@ -32,8 +32,8 @@ export default function PersistentClearAllButton(): JSXInternal.Element {
 
   const renderClearAllButton = () => {
     const clearAllText = localize("PersistentClearAllButton.clearAll", "Clear All");
-
-    return selection[type] === "search" || selection[type] === "fileHistory" ? (
+    const selectionType = ["search", "fileHistory", "encodingHistory"];
+    return selectionType.includes(selection[type]) ? (
       <VSCodeButton title={clearAllText} appearance="secondary" style={{ maxWidth: "20vw", marginRight: "15px" }} onClick={handleClick}>
         Clear All
       </VSCodeButton>