+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentToolBar/PersistentDropdownOptions.tsx
@@ -34,10 +34,17 @@ export default function PersistentDropdownOptions({ handleChange }: Readonly<{ h
     </VSCodeOption>,
   ].filter((option) => dataPanelContext.type === "ds" || option.props.value !== "dsTemplates");
 
+  const optionsEncodingHistory = [
+    <VSCodeOption value="encodingHistory" key="encodingHistory">
+      Encoding History
+    </VSCodeOption>,
+  ].filter((option) => dataPanelContext.type === "uss" || dataPanelContext.type === "ds" || option.props.value !== "encodingHistory");
+
   return (
     <div style={{ display: "flex", flexDirection: "row", alignItems: "center", margin: "15px 15px 15px 0px" }}>
       <VSCodeDropdown id="dropdown-persistent-items" style={{ maxWidth: "20vw" }} onChange={(event: any) => handleChange(event.target.value)}>
         {options}
+        {optionsEncodingHistory}
       </VSCodeDropdown>
     </div>
   );